Separated into multiple packages
Nucleus-API importing in Util, Test, and Entity
