+++ src/main/kotlin/com/immanuelqrw/core/api/controller/ReadController.kt
@@ -1,8 +1,8 @@
 package com.immanuelqrw.core.api.controller
 
 import com.immanuelqrw.core.api.Getable
-import com.immanuelqrw.core.api.model.BaseEntity
 import com.immanuelqrw.core.api.service.BaseService
+import com.immanuelqrw.core.entity.BaseEntity
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
