+++ src/main/kotlin/com/immanuelqrw/core/api/controller/WriteController.kt
@@ -1,8 +1,8 @@
 package com.immanuelqrw.core.api.controller
 
 import com.immanuelqrw.core.api.Writable
-import com.immanuelqrw.core.api.model.BaseEntity
 import com.immanuelqrw.core.api.service.BaseService
+import com.immanuelqrw.core.entity.BaseEntity
 import org.springframework.beans.factory.annotation.Autowired
 
 /**
