--- src/main/kotlin/com/immanuelqrw/core/api/model/Entityable.kt
@@ -1,6 +1,0 @@
-package com.immanuelqrw.core.api.model
-
-/**
- * Interface for entities
- */
-interface Entityable : Identifiable, Entryable
