--- src/main/kotlin/com/immanuelqrw/core/api/model/Entryable.kt
@@ -1,16 +1,0 @@
-package com.immanuelqrw.core.api.model
-
-import java.time.LocalDateTime
-
-/**
- * Interface specifying required fields needed to be an entry
- *
- * @property createdOn When entity's entry was created
- * @property modifiedOn When entity's entry was last modified
- * @property removedOn When entity's entry was removed
- */
-interface Entryable {
-    val createdOn: LocalDateTime?
-    val modifiedOn: LocalDateTime?
-    val removedOn: LocalDateTime?
-}
