--- src/main/kotlin/com/immanuelqrw/core/api/model/Identifiable.kt
@@ -1,10 +1,0 @@
-package com.immanuelqrw.core.api.model
-
-/**
- * Interface specifying required field to be identifiable
- *
- * @property id Unique identifier for entity
- */
-interface Identifiable {
-    val id: Long?
-}
