+++ src/main/kotlin/com/immanuelqrw/core/api/repository/BaseRepository.kt
@@ -1,6 +1,6 @@
 package com.immanuelqrw.core.api.repository
 
-import com.immanuelqrw.core.api.model.BaseEntity
+import com.immanuelqrw.core.entity.BaseEntity
 import org.springframework.data.jpa.repository.support.JpaRepositoryImplementation
 
 /**
