+++ src/main/kotlin/com/immanuelqrw/core/api/service/SearchService.kt
@@ -1,8 +1,8 @@
 package com.immanuelqrw.core.api.service
 
 import com.immanuelqrw.core.api.filter.SearchSpecificationsBuilder
-import com.immanuelqrw.core.api.model.BaseEntity
 import com.immanuelqrw.core.api.utility.Utility.SEARCH_PATTERN
+import com.immanuelqrw.core.entity.BaseEntity
 import org.springframework.data.jpa.domain.Specification
 import java.util.regex.Pattern
