+++ src/main/kotlin/com/immanuelqrw/core/api/test/example/Solid.kt
@@ -1,14 +1,8 @@
 package com.immanuelqrw.core.api.test.example
 
-import com.immanuelqrw.core.api.model.BaseEntity
-import com.immanuelqrw.core.api.utility.DateTimeFormatter
-import org.hibernate.annotations.CreationTimestamp
-import org.hibernate.annotations.UpdateTimestamp
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
-import org.springframework.format.annotation.DateTimeFormat
-import java.time.LocalDateTime
-import javax.persistence.*
+import com.immanuelqrw.core.entity.BaseEntity
+import javax.persistence.Entity
+import javax.persistence.Table
 
 /**
  * Solid entity class
