+++ src/main/kotlin/com/immanuelqrw/core/api/test/example/SolidController.kt
@@ -1,9 +1,6 @@
 package com.immanuelqrw.core.api.test.example
 
 import com.immanuelqrw.core.api.controller.BaseController
-import com.immanuelqrw.core.api.model.BaseEntity
-import com.immanuelqrw.core.api.service.BaseService
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.*
 
 /**
