+++ src/main/kotlin/com/immanuelqrw/core/api/test/example/SolidSearchSpecificationsBuilder.kt
@@ -2,4 +2,4 @@ package com.immanuelqrw.core.api.test.example
 
 import com.immanuelqrw.core.api.filter.SearchSpecificationsBuilder
 
-open class SolidSearchSpecificationsBuilder: SearchSpecificationsBuilder<Solid>()
+class SolidSearchSpecificationsBuilder: SearchSpecificationsBuilder<Solid>()
