--- src/main/kotlin/com/immanuelqrw/core/api/utility/DateTimeFormatter.kt
@@ -1,13 +1,0 @@
-package com.immanuelqrw.core.api.utility
-
-// - Move into core.utility package
-/**
- * Configuration object which stores DateTime Formatting configuration values
- */
-object DateTimeFormatter {
-    /**
-     * Date Formatting constants
-     */
-    const val DATE_TIME_PATTERN: String = "YYYY-MM-DDThh:mm:ssZ"
-    const val DATE_PATTERN: String = "YYYY-MM-DD"
-}
